lambda-code/requirements_mecab.txt
boto3>=1.34.0
urllib3>=2.0.0
mecab-python3>=1.0.6
